Implementation of Creating Students in a Class
Teachers can now create students! This was a bit tricky to do with the auth, because for whatever reason Firebase automatically switches auth to the student's auth whenever you create a new user. Had to get a little creative with this by getting the teacher's auth again by having them enter their password to confirm. It's not the most elegant solution, but I thought it was pretty clever!