Update to Teacher Dashboard and Word List
This push essentially just allows teachers to now add new classes to their dashboard (in addition to students). Additionally, the word lists are now separated on the word list screen so that each list is in a collapsable menu. This way, the UI looks much cleaner and doesn't list all of them out in an unclear fashion.